refactor list_devices
Each platform specific implementation is now in its own translation unit: linux_list_devices, macos_list_devices, win32_list_devices.
This is in preparation for implementing the macos_list_devices parsing of the swift script stdout
