+++ video/CMakeLists.txt
@@ -12,8 +12,9 @@ configure_file(${CMAKE_CURRENT_SOURCE_DIR}/config.h.in ${CMAKE_CURRENT_BINARY_DI
 
 # Command line executable with a simple OpenCV GUI
 add_executable(video video_main.cpp run_tests.cpp run_application.cpp
-               list_devices.cpp opencv_ui.cpp
-               input_device.cpp opencv_window.cpp)
+               opencv_ui.cpp list_devices.cpp win32_list_devices.cpp
+               linux_list_devices.cpp macos_list_devices.cpp input_device.cpp
+               opencv_window.cpp)
 
 # Open Computer Vision (github.com/opencv/opencv)
 find_package(OpenCV REQUIRED)
