+++ video/run_application.h
@@ -1,5 +1,7 @@
 #pragma once
 
 namespace frank::video {
+
 [[noreturn]] extern void run_application(int argc, char const *argv[]);
+
 }
