+++ video/run_tests.h
@@ -1,5 +1,7 @@
 #pragma once
 
 namespace frank::video {
+
 [[noreturn]] extern void run_tests(int argc, char const *argv[]);
+
 }
